requirements.txt
aiogram==3.11.0
asyncio==3.4.3
pymongo==4.8.0
motor==3.5.1
asyncio==3.4.3
pytest==8.3.2
pytest-asyncio==0.23.8
pyyaml==6.0.2
python-dotenv==1.0.1